Add new API without AAD masking
The AAD mask changed for new packet types added after the
BT 4.2 version.
All new masks have masked out more bits than before,
so it is possible to keep using the old API, but it can
be confusing for developers to have the mask applied in 2
places. Let's add a new API for which the mask is applied
by the user.

Fixed a few typos also.

Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>